Web/backend/requirements.txt
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.1.2
billiard==4.2.0
black==23.11.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.4.0
cryptography==41.0.7
daphne==4.0.0
Django==4.2.9
django-cors-headers==4.3.1
django-extensions==3.2.3
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.1.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.4
MarkupSafe==2.1.3
msgpack==1.0.8
mypy-extensions==1.0.0
neat-python==0.92
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
twisted==23.10.0
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
WebOb==1.8.7
werkzeug==3.0.1
zope.interface==6.1
web3
py-solc-x
eth_tester